Fix ShellCheck linter warnings on `build-darwin.sh`
This patch fixes the warnings that our ShellCheck linter catches on
the `build-darwin.sh` file.  Most of these changes are about quoting
expanded variables, which in our case is generally good: for example,
we want to correctly handle the case of a user’s build directory
having a space in its full path.  In the case of `CMAKE_OPTIONS`, we
do this by changing the string variable into an array of strings we
can correctly quote, both when creating the array and when expanding
it as arguments.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>