Fix ShellCheck linter warnings on `build-ubuntu.sh`
This patch fixes the warnings that our ShellCheck linter catches on
the `build-ubuntu.sh` file.  As with the previous patch, the majority
of these changes are about quoting expanded variables.  Here
additionally, we are using a HereDoc as multiline comment for the user
to copy and paste, so we disable a linter warning `SC2188` about
[redirection missing a command][sc2188] for this line.

[sc2188]: https://www.shellcheck.net/wiki/SC2188

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>