Make `bmqa_mocksession` match `bmqa_abstractsession` in `const`ness
The `expect_XYZ` functions in `bmqa_mocksession` (e.g.,
`expect_configureQueueSync`) are run using the `BMQA_EXPECT_CALL`
macro from a mock, and are meant to parallel the functions
`XYZ` (e.g., `configureQueueSync`) in `bmqa_abstractsession`.  At the
moment, several of these functions `expect_XYZ` differ from their
corresponding `XYZ` functions in the `const`ness of their parameters:

  * `expect_closeQueueSync(QueueId*, ...)`, but
    `closeQueueSync(const QueueId*, ...)`

  * `expect_closeQueueAsync(QueueId*, ...)`, but
    `closeQueueAsync(const QueueId*, ...)`

  * `expect_configureQueueSync(QueueId*, ...)`, but
    `configureQueueSync(const QueueId*, ...)`

As a result, it is not possible to use `BMQA_EXPECT_CALL` from a mock
session that implements the `AbstractSession` interface.

This patch changes the arguments to each `expect_XYZ` functions to
take the same cv-qualifiers as their corresponding arguments in each
`XYZ` function.

Fixes: #96
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>