Add `DIR_INSTALL` variable to `bin/build-darwin.sh`
This patch adds a new `DIR_INSTALL` environment variable that the user
can set to where they want BlazingMQ and its dependencies to be
installed to.  By default, this variable is set to `DIR_ROOT`, namely
the root of the source tree.

In the process, this fixes an issue with `CMAKE_INSTALL_LIBDIR`, which
is meant to be a path relative to `CMAKE_INSTALL_PREFIX`.  Previously,
the build script would generate a CMakeCache that tries to install
libraries to `${DIR_ROOT}/${DIR_ROOT}/lib`.  These changes correct
this behavior.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>