Clarify license notice for `Just the Docs`
The previous text within the license section for our documentation
site was unclear, simply referencing that we use Just the Docs.  This
patch clarifies that Just the Docs is the Jekyll documentation theme
we use, that it is licensed under the MIT license, and links to the
Just the Docs license file, which contains the copyright and
permission notices that we need to distribute with the software
according to the MIT license.  None of these changes affect our
license obligations, but clarify that we are in fact following our
obligations.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>