Move Just the Docs license to `licenses` folder
Our current guidance is to store the licenses of our dependencies
within the directory `licenses`.  The license for Just the Docs was
originally stored with the documentation itself, but should be stored
with all our other dependencies.  This patch moves the license into
the directory with all our other dependencies’ licenses.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>