#107 Fixup README for Just the Docs documentation website
* Clarify license notice for `Just the Docs`

The previous text within the license section for our documentation
site was unclear, simply referencing that we use Just the Docs.  This
patch clarifies that Just the Docs is the Jekyll documentation theme
we use, that it is licensed under the MIT license, and links to the
Just the Docs license file, which contains the copyright and
permission notices that we need to distribute with the software
according to the MIT license.  None of these changes affect our
license obligations, but clarify that we are in fact following our
obligations.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* Improve documentation for local build on Mac

This patch removes some stray, unclear references and replaces them
with clearer instructions for users to build the documentation
locally.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

* Move Just the Docs license to `licenses` folder

Our current guidance is to store the licenses of our dependencies
within the directory `licenses`.  The license for Just the Docs was
originally stored with the documentation itself, but should be stored
with all our other dependencies.  This patch moves the license into
the directory with all our other dependencies’ licenses.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>

---------

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>