Remove public BDE-style package documentation
This patch removes the public BDE-style package documentation for our
three public packages: `bmqa`, `bmqpi`, and `bmqt`.  This
documentation is old and incomplete, and a more update-to-date version
can be found already in the `README.md` for each of these packages.
These `README.md` files can be used directly by Doxygen as well,
rather than the Doxygen component documentation being generated as a
separate step from the BDE-style documentation.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>