docs: Prevent Jekyll from mangling our Sphinx documentation
The Python SDK API documentation is generated with Sphinx, which creates a
directory structure that is incompatible with the Jekyll job that GitHub uses
to build the rest of our documentation.  Because the API docs have already been
pregenerated, we want to make GitHub serve the contents of the Python SDK API
documentation directory directly.  This patch adds a `.nojekyll` file that does
that.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>