docs: Teach Jekyll to directly include generated Python docs
The previous patch tried to do this with a `.nojekyll` file, but this file only
works when placed in the root of the GitHub pages directory, and serves to turn
off any Jekyll generation for the entire site.  We do want Jekyll to generate
most of our site for us, but we don’t want it to ignore directories that the
Python documentation generator Sphinx creates for us which begin with
underscores.  This patch reverts the `.nojekyll` attempt, and instead modifies
the Jekyll configuration to explicitly include these directories.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>