docs: Add a new Doxygen alias for referencing BlazingMQ components
All BlazingMQ components live within the namespace `BloombergLP`, under
which package namespaces `bmqa`, `bmqp`, etc. live.  In most places in
our current documentation, though, this qualifier is elided: the class
`BloombergLP::bmqa::AbstractSession` is referred to only as
`bmqa::AbstractSession`.  While this is usually not an issue, if we want
Doxygen to generate links to the documentation for these symbols, we
must prefix them with `BloombergLP::`.  This gives us two options:

  1. Any time we reference a symbol in our documentation, we must make
     sure to that we reference its fully-qualified name, including the
     `BloombergLP::` namespace prefix.  These changes would take the
     form

         … 'bmqa::AbstractSession' …
     →
         … @ref BloombergLP::bmqa::AbstractSession ….

     This will make our documentation more noisy and harder to read.

  2. We add our own Doxygen command alias that takes a symbol name and
     generates the above.  We will have to make sure to use this alias
     whenever we modify our Doxygen documentation, or symbol references
     will not work correctly.  Then, changes would take the form

         … 'bmqa::AbstractSession' …
     →
         … @bbref{bmqa::AbstractSession} …,

     which is only a little noisier than what we have now.

This patch takes the second approach, and introduces a Doxygen command
alias named `bbref`, which performs the substition described above.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>