docs: Convert `bmqa` component documentation to Doxygen style
During the open-sourcing process, we converted all class and function
documentation over from the BDE documentation style to Doxygen +
Markdown using an automated tool.  This tool did not affect
package-group-level documentation, package-level documentation, or
component-level documentation, though, which is still in the previous
BDE style, and which is thus not picked up when we run Doxygen.  This is
a big problem, since much of our SDK’s documentation, as well as
examples, are given in component-level documentation rather than on the
concrete classes, functions, or macros.  This also means we cannot run
Doxygen on our source code to produce updated versions of the
documentation when we release, meaning our documentation is still stuck
at where it was with our first open-source release.

This patch performs the first big step towards rectifying this
situation: it converts the old BDE-style component-level documentation
into Doxygen + Markdown documentation, applied to the header file
associated with each component.  After applying this patch, Doxygen is
able to pick up the component-level documentation and to produce
formatted pages for each component.

Documenting components in Doxygen
=================================

Doxygen has no native support for BDE-style organization, with package
groups, packages, and components, so we need to map these organizational
concepts onto Doxygen’s more impoverished and concrete ontology.  There
are two rough ways of going about this, described below.

The first strategy is the one that we employed before converting the
documentation on classes, functions, and macros from BDE-style to
Doxygen + Markdown format: use the `bdedox` tool from the
[bde-tools][bde-tools] repository, which to runs a filter program over
each package group, package, and component in the repository.  Using the
documentation embedded within a BDE-style repository, this filter
produces temporary Markdown files package groups, packages, and
components.  Those temporary Markdown files then serve as additional
input to Doxygen, rendering as separate Doxygen _pages_—free-form
documentation files that Doxygen imposes no structure or hierarchy on.
`bdedox`, however, expects all documentation to be in BDE-style,
including class- and function-level documentation; since we converted
both of these into Doxygen style, we cannot use the tool.  The only way
to follow this method of teaching Doxygen to treat our component-level
documentation as pages is to write our own filter program.  This is a
larger undertaking to support a legacy documentation format, which we
have already half-converted away from.

The second and alternative strategy is to map package groups, packages,
and components onto entities that Doxygen already natively understands.
We elected to use this strategy for this patch.  However, in the case of
components, there is no obvious C++-level entity that uniquely
corresponds to a component.  Instead, in this patch, we commit the sin
of conflating the header file associated with a component with the
component itself.  That is, we apply _component-level documentation_ to
the _component header file_.  After applying this patch, then, the
header file within a component (say, `bmqa_component`) will have a
Doxygen-style comment near its top that starts with `@file
bmqa_component.h`, and which then contains the documentation for
`bmqa_component`.

This approach, though easier, has its downsides.

  1. First, when referencing the documentation for another component in
     the library, we have to reference the header file name rather than
     the component name.  So, if some documentation comment wants to
     point to the documentation for component `bmqa_component`, it must
     reference it with `@ref bmqa_component.h`.  This might be
     confusing to users familiar with our BDE-style organization, but
     if we prefer, we can render it correctly with
     `\[bmqa_component\](@ref bmqa_component.h)`, at the cost of
     noisier documentation comments.

  2. Second, when generated, component-level documentation will be found
     under the `Files` documentation in the left-hand navigation column,
     buried within several subdirectories.  We can partially solve this
     by configuring Doxygen to strip the unnecessary subdirectories from
     file paths, but users may still not be inclined to look under
     `Files` for important documentation.  I cannot see a way to rename
     this portion of the documentation from `Files` to `Components`.

[bde-tools]: https://github.com/bloomberg/bde-tools

Notes on Documentation Markup
=============================

Doxygen understands Markdown and HTML formatting rather than BDE-style
documentation, so we have manually converted BDE-style markup to
Markdown and, in more advanced cases, HTML.  The resulting HTML is
less-readable than BDE-style markup, but produces easier-to-understand
Doxygen documentation.

Sections within the component-level documentation are rendered using
Markdown headings, with unique IDs, constructed by prefixing the
component name with a short identifier.  Unfortunately, even if they are
not referenced from other components, these IDs must be unique across
the entire project, meaning we cannot get by with shorter identifiers.
Doxygen will warn us when it encounters an identifier that is not
unique, though, which is a good way to catch these issues.  Within the
same documentation comment, we can link to a section just using a simple
Markdown link: `\[some link\](#bmqa_component_section)`.  If we want to
link to a section in a different component, we can use Doxygen’s
referencing capability: `\[some link\](@ref #bmqa_component_section)`.

The biggest benefit of Doxygen is that we do not need to list the
entities that are exposed by some component manually in the
documentation comment; instead, Doxygen will collect these itself and
produce the list for us.  As such, we have removed the `@CLASSES`
section from the component-level documentation.  At the moment, this is
not a perfect substitution, as some private entities witehin components
are not properly marked as private, and are thus included within the
generated entity lists.  We will fix this with a later patch.

We have mapped BDE `@PURPOSE` sections to Doxygen `@brief` documentation
and `@DESCRIPTION` sections to Doxygen long documentation (which isn’t
explicitly marked in a Doxygen documentation comment).  We have also
used the Doxygen `@see` command to implement BDE `@SEE_ALSO` sections.

Testing
=======

The changes in this patch are simple in principle, but were performed
manually.  Verifying that the changes in this patch are correct was done
by generating the documentation by running `doxygen` from the root
directory of this repository, ensuring that no new warnings or errors
are produced by it, and then manually reviewing the generated
documentation for its fidelity.  This last step is tedious and prone to
mistakes, so we relied heavily on the warnings from Doxygen to ensure
our documentation was correct.

Despite this, there are still a few warnings that are produced.  Doxygen’s
parser seems to be confused by several BDE macros, which it cannot find the
definitions of.  These warnings can be fixed by predefining those macros, but
this is left to a future patch.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>