docs: Strip `src/groups/bmq` from component paths
Now that component-level documentation has been conflated with header
file-level documentation, the component-level documentation is buried
within the `Files` section in the left-hand navigation pane of our
generated Doxygen documentation.  Several levels of hierarchy within
this output are noise: namely, we do not need the `src/groups/bmq`
directories that are incidental to the logical hierarchy of our
components.

This patch teaches Doxygen to strip out those directories from the names
of files it searches, yielding a documentation structure that is closer
to the BDE-style package and component hierarchy.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>