docs: Expand macros that confuse Doxygen’s parser
Our Doxygen is not configured to search through BDE’s header files for
macros to expand, so they are passed, unexpanded, on to Doxygen’s
parser.  In certain cases, this confuses the parser: for instance, the
lack of a semicolon after `BSLMF_NESTED_TRAIT_DECLARATION(..)‘ uses
results in an apparent syntax error, which causes Doxygen to build our
documentation with warnings.  Other BDE macros that are meant to enable
C++11 and later features to be used only when available, like
`BSLS_KEYWORD_FINAL`, are also unexpanded, and result in Doxygen failing
to properly document symbols to which they are attached.

This patch teaches Doxygen to ignore or properly expand the above two
macros, removing several warnings from our documentation builds.  There
may be more macros that need to be added, but these are the only ones
for which Doxygen issued warnings.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>