sdk: Fix clang `-Wmismatched-tags` warning
The `bmqa::MessageImpl` class is declared in the `bmqa_message` component as a
`struct`, but is forward-declared in `bmqt_subscription` as a `class`.  This is
valid C++ with our desired semantics, but the clang still warns that under the
Microsoft C++ ABI, this may result in linker errors.  Although we do not
support Windows as a target platform, this warning is pernitious: many
components that use `bmqt_subscription` will see this warning; including those
of users!

While it would be easy to turn off this warning, this patch instead fixes the
warning by changing the `class` forward declaration in `bmqt_subscription` to
match the `struct` declaration in `bmqa_message`.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>