build: Do not pass `-Wlogical-ops` to clang
The supported compiler on modern MacOS systems is clang; the `gcc` command
forwards to `clang` underlyingly, unless a user explicitly installs GCC.
Unfortunately, our build system was assuming GCC as the compiler on MacOS
systems.  For the most part, this had no adverse effects, as the two frontends
attempt to be compatible.  However, there is one warning that GCC supports that
clang does not, which we were inadvertently passing: `-Wlogical-op`, which in
clang is superseded by `-Wlogical-op-parentheses`.

The primary issue here is that the we outsource some of our CMake configuration
to BDE Tools, but we need to manually select which platform to use.  This patch
manually selects `clang` rather than `gcc` in the `bin/build-darwin.sh` script,
which silences the above warning.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>