docs: Use Doxygen `@deprecated` tags and `BSLA_DEPRECATED`
This patch converts the unstructured, artisanal API deprecation warnings
in the documentation for MWC and for our public SDK (`bmqa`, `bmqt`)
into Doxygen `@deprecated` tags, and also adds corresponding
`BSLA_DEPRECATED` attributes to the affected entities.

First, by converting deprecation warnings into tags that Doxygen knows
about, Doxygen is able to construct a list of deprecated entities in the
library, as well as gray them out in generated documentation for the
classes or components they appear in.  The changes here consist of
converting handwritten `DEPRECATED:` sections into `@deprecated`
paragraphs, and correcting minor grammatical errors in those paragraphs.

Second, introducing `@deprecated` tags into the Doxygen documentation
comments causes clang to issue warnings that there is no corresponding
`[[deprecated]]` annotation on the entity that the documentation comment
applies to.  Thus, for each modified Doxygen documentation comment, we
also add the `BSLA_DEPRECATED` annotation, which expands to
`[[deprecated]]` when compiled with an implementation that supports this
attribute.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>