build: Disable `-Wundef` on clang
The BDE idiom for using `BSLS_COMPILERFEATURES_SIMULATE_CPP11_FEATURES`
is by using it in the condition of a preprocessor conditional `#if`, as
follows:

    #if !BSLS_COMPILERFEATURES_SIMULATE_CPP11_FEATURES

When compiling on a platform that supports C++11, this macro is
undefined; the behavior in this situation is well-defined, though, and
the macro is replaced with `0`, rendering the above condition true.

Clang, however, warns in this situation, letting us know that we are
expanding a macro that is undefined to `0`, which we may not want to do.
Because of how commonly we use this idiom in our codebase, this patch
turns off this warning across all files.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>