docs: Exclude internal namespaces
In several places in our SDK, we need to forward declare a symbol from
one of the implementation packages (e.g., `bmqimp`).  These packages do
not form part of the public API of `libbmq`, and we do not want to
include them in the documentation.  Unfortunately, it looks like the
best way to do this is to manually and explicitly exclude the namespaces
for these symbols in the `Doxyfile`.  This patch excludes the namespaces
of components outside of `bmqa`, `bmqpi`, and `bmqt` that we need to
forward-declare, including private `libbmq` packages and BDE packages.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>