Lock version of BDE in `build-ubuntu.sh`
The `build-darwin.sh` script builds its dependencies BDE and ntf-core
from tagged releases available in their git repositories: for BDE,
version 3.117.0.0, released on 21 Apr 2023, and for ntf-core, the
latest version tag.  Similarly, the `build_deps.sh` script that is
used to build our Docker image provides a `fetch_git` function that
allows us to build from a particular git tag as well; we call it with
these same versions.  `build-ubuntu.sh`, on the other hand, always
checks out from `main`.  In particular in the case of `ntf-core`,
`main` is not always guaranteed to build successfully, meaning we can
get spurious build failures.

This patch copies the `git clone` commands from `build-darwin.sh` to
`build-ubuntu.sh`, ensuring that we’re always locked in to a released
version for BDE and ntf-core when building on Linux.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>