Bump BDE version in build scripts
This patch bumps the version of BDE that we clone in our build scripts
from version 3.117.0.0, relased 21 Apr 2023, to version 4.8.0.0,
released 26 Apr 2024.

In the process of updating BDE to a later version that supports C++20,
we discovered an issue with our Ubuntu and Docker build scripts, which
were building BDE and BlazingMQ with C++17, but ntf-core without an
explicit C++ version.  Before BDE and bde-tools knew about C++20, this
was not an issue, but with a newer BDE, ntf-core is being built with
C++20 support in BDE.  Then, when linking the C++17 BDE and the C++20
ntf-core in BlazingMQ yields an error:

    /usr/bin/ld: /opt/bb/lib64/opt_dbg_exc_mt/libnts.a(ntsscm_version.cpp.o):(coercion+0x0): undefined reference to `BloombergLP::bsls_libraryfeatures_CPP20_ABI'

The solution to this was already present in `build-darwin.sh`, which
builds ntf-core using C++17, by configuring it with the flag `--ufid
opt_64_cpp17`.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>