Teach `build_deps` not to clean ntf before rebuilding
ntf-core’s `configure` script automatically cleans build artifacts
before it runs.  Perhaps this is a good default, but because we want
to cache our build artifacts in GitHub actions, we don’t want this.
This patch passes the `--keep` flag to `configure` when we call it
from `build_deps.sh`, which preserves existing build artifacts each
time we configure.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>