Remove website files
This patch removes the old website files from our `main` branch, except
for those assets that are also used by our Doxygen documentation
generation.  Because these files now live on the `gh-pages` branch,
having an older, out-of-date copy of these files in `main` is
confusing.

This patch also fixes up the `build.yaml` documentation step, which
verifies that the docs can be built without error.  This step will
eventually be removed in favor of a more fleshed-out
`documentation.yaml`, but we will keep it for the moment.

Testing
========================================================================

The documentation still builds correctly after applying this patch.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>