Fix[bmqa]: Remove unused `Session::impl()` function
This function is not used within the BlazingMQ code, and exposes
publically the PIMPL implementation that is used by `bmqa::Session`.
Because we do not use this function, and because it is part of the
public (documented!) API, users could break our session invariants.
This patch removes this function from the API.  While this is a breaking
change, we do not want any user to be using this function, so we should
be okay with breaking them.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>