Migrate mqba docs into a Doxygen file
This patch removes the old `doc/*.txt` file and migrates some
information in it into a file Doxygen can read.  The information not
migrated is:

  1. the table of contents for this package, which will be generated by
     Doxygen automatically, and kept in-sync with any documentation
     changes made in the files, and

  2. the hierarchical synopsis, which is out-of-date.

These are the same changes that we have already made in the public
documentation, as in commits bc20ab002a101147bef682bad076e9329f7a31fe
and f3ca1d24138db58b4eea57e6a74872117200bb56.  See in particular the
rational in the latter’s commit message.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>