Add black as a dependency for Python code
This patch adds the current `HEAD` of `main` for the Python formatter
`black` to our dependencies.  `black` used to be a transitive
dependency, but this was removed as part of
b39062018c2ef3e6b793aef8fb0e217401bbc642.  The decision to use `HEAD`
instead of a particular version is only to match the version used in CI
in `.github/formatting-check.yaml`; if we elect to use a particular
version, we should do so there as well.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>