Rename auto-subscriptions to application subscriptions
This patch renames all instances of “auto subscriptions” in the code to
“application subscriptions”, which is what we use externally to describe this
feature.  Because application subscriptions are only evaluated and referenced
in the broker, there is no chance of an API-breaking change.  Indeed, only the
`mqb` package group is affected.

This patch leaves log messages that reference auto subscriptions;
this will be handled separately, as tools that parse the log for lines about
application subscriptions may depend on this older output.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>