Fix[MQB]: Select active node quickly when no cluster node in same DC
When a broker connects to a cluster and attempts to select an active
node, the broker attempts to select a node in the same data center.
After 10 seconds, if no active node has been selected, the broker
relaxes the data center requirement and picks any available node. If a
cluster does not have any nodes in the broker's data center, this leads
to a 10 second delay on the first open queue request to a cluster.

This commit changes the active node selection logic for the case where
there is no cluster node within the same data center as a broker. In
this case, the broker will ignore the data center may use any available
node as the active node.

Signed-off-by: Christopher Beard <cbeard9@bloomberg.net>