CI: Publish container images to GHCR.io
We had tested publishing BlazingMQ images to DockerHub, but we believe
the GitHub Container Registry (GHCR.io) to be a better platform for
our needs.  This patch updates the `publish` workflow to publish our
Docker image publicly to GHCR.io every time we tag a release, and on a
manual dispatch.  The image is built both for x64 and arm64
architectures, and is automatically tagged with the version number
from the Git tag, `latest` if it’s the most recent release, or as
`main` if the workflow was manually dispatched from the `main` branch.

Closes: #628
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>