CI: Create GitHub (pre)release on pushed version tag
This patch extends our Publish workflow to create a GitHub prerelease when a
release candidate tag of the form `v(.*)rc(.*)` is pushed, and a GitHub release
when any other release tag of the form `v(.*)` is pushed.  In either case, we
make the prerelease or release a draft, so we can manually write in a release
description, in addition to the automatically generated release notes.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>